pyproject.toml
[project]
name = "pensarsecurityeval"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.49.0",
    "black>=25.1.0",
    "openai>=1.73.0",
    "pandas>=2.2.3",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "tqdm>=4.67.1",
]
